requirements.txt
pyo
pyimgui
moderngl
moderngl_window
rich
numpy
pyzmq
